Move the Mesa LLVM hack from GLES to Vulkan.
The Vulkan driver has the exact same issue, we've just been relying on
GLES being loaded first.
